Fixup decorator handling
